feat: remove react-app-env.d.ts file
This commit removes the react-app-env.d.ts file, which was previously located in the src directory. The file mode was set to deleted and its contents were completely removed.

This file was previously used to declare global types for the React application, but it is no longer necessary as these types are now included in the default configuration of create-react-app.

This change helps to simplify the project structure and reduce unnecessary code, improving maintainability and reducing the risk of errors.